+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 5.6.1 (in progress)
 
-* Your contribution here!
+##### Bug fixes / Improvements
+* [#1562](https://github.com/oshi/oshi/pull/1562): Fix missing space in WMI process query - [@dbwiddis](https://github.com/dbwiddis).
+* [#1566](https://github.com/oshi/oshi/pull/1566): Handle new WinAPI Logical Processor Information types - [@dbwiddis](https://github.com/dbwiddis).
 
 # 5.6.0 (2021-03-01)