+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [#1566](https://github.com/oshi/oshi/pull/1566): Handle new WinAPI Logical Processor Information types - [@dbwiddis](https://github.com/dbwiddis).
 * [#1567](https://github.com/oshi/oshi/pull/1567): Handle empty process performance registry query - [@dbwiddis](https://github.com/dbwiddis).
 * [#1569](https://github.com/oshi/oshi/pull/1569): Fix udev reference leak in LinuxNetworkIF - [@dbwiddis](https://github.com/dbwiddis).
+* [#1576](https://github.com/oshi/oshi/pull/1576): JNA 5.8.0 - [@dbwiddis](https://github.com/dbwiddis).
 
 # 5.6.0 (2021-03-01)